Utilize separate compilation which generates a library which links with the program
Each source file will generate its own object file, which gets linked into "cligCore.o". 

cligCore.o now has to be linked during compilation to the source program.

The user now includes the cligCore header only in their code.
