+++ .gitignore
@@ -32,5 +32,7 @@
 *.app
 
 # .vscode files
-*.vscode/c_cpp_properties.json
-*.vscode/settings.json
+.vscode/c_cpp_properties.json
+.vscode/settings.json
+build/
+bin/
