+++ src/cligCore/console/_debuggingTests.cpp
@@ -1,3 +1,4 @@
+#include "clig.hpp"
 namespace cligCore {
   namespace console {
     void debug() {
