+++ src/cligCore/console/getConsoleHeight.cpp
@@ -1,21 +1,25 @@
+#include "clig.hpp"
+
 namespace cligCore {
   namespace console {
-    // Used to get the current Width of the console window
-    int getConsoleWidth() {
+    // Used to get the current Height of the console window
+    int getConsoleHeight() {
 #if defined( _WIN32 ) || defined( _WIN64 )
-
+#include "windows.h"
+#include <limits.h>
       CONSOLE_SCREEN_BUFFER_INFO csbi;
-      int columns;
+      int rows;
       GetConsoleScreenBufferInfo( GetStdHandle( STD_OUTPUT_HANDLE ), &csbi );
-      columns = csbi.srWindow.Right - csbi.srWindow.Left + 1;
-
-      return columns;
-
+      rows = csbi.srWindow.Bottom - csbi.srWindow.Top + 1;
+      return rows;
 #elif defined( __unix__ ) || defined( __unix )
-
+#include <sys/ioctl.h>
+#include <termios.h>
+#include <cstdlib>
+#include <unistd.h>
       struct winsize max;
       ioctl( 0, TIOCGWINSZ, &max );
-      return max.ws_col;
+      return max.ws.col
 #endif
     }
   } // namespace console
