+++ src/cligCore/console/getConsoleWidth.cpp
@@ -1,23 +1,23 @@
+#include "clig.hpp"
+
 namespace cligCore {
   namespace console {
-    // Used to get the current Height of the console window
-    int getConsoleHeight() {
+    // Used to get the current Width of the console window
+    int getConsoleWidth() {
 #if defined( _WIN32 ) || defined( _WIN64 )
-#include "windows.h"
-#include <limits.h>
+
       CONSOLE_SCREEN_BUFFER_INFO csbi;
-      int rows;
+      int columns;
       GetConsoleScreenBufferInfo( GetStdHandle( STD_OUTPUT_HANDLE ), &csbi );
-      rows = csbi.srWindow.Bottom - csbi.srWindow.Top + 1;
-      return rows;
+      columns = csbi.srWindow.Right - csbi.srWindow.Left + 1;
+
+      return columns;
+
 #elif defined( __unix__ ) || defined( __unix )
-#include <sys/ioctl.h>
-#include <termios.h>
-#include <cstdlib>
-#include <unistd.h>
+
       struct winsize max;
       ioctl( 0, TIOCGWINSZ, &max );
-      return max.ws.col
+      return max.ws_col;
 #endif
     }
   } // namespace console
