+++ src/cligCore/input/getKeyInput.cpp
@@ -1,3 +1,4 @@
+#include "clig.hpp"
 namespace cligCore {
   namespace input {
     Keys getKeyInput() {
