--- src/console/_console.cpp
@@ -1,6 +1,0 @@
-#include "clear.cpp"
-#include "getConsoleHeight.cpp"
-#include "getConsoleWidth.cpp"
-#include "getPassword.cpp"
-#include "toggleEcho.cpp"
-#include "_debuggingTests.cpp"
