--- src/input/_input.cpp
@@ -1,1 +1,0 @@
-#include "getKeyInput.cpp"
