+++ src/main.cpp
@@ -1,11 +1,12 @@
+#include "clig.hpp"
+
 #include <iostream>
 #include <string>
 #include <vector>
-#include "termcolor.hpp"
-#include "clig.cpp"
 
 int main() {
   std::ios_base::sync_with_stdio( false ); // Reduces flashing, but introduces other graphical oddities.
+
   std::vector<std::string> menuContent{
       "1. Console", "2. Controls", "3. Filesystem", "4. Input", "5. Types", "6. Quit",
   };
