--- src/types/types.cpp
@@ -1,2 +1,0 @@
-#include "Range/Range.cpp"
-#include "Menu/Menu.cpp"
